+++ lib/stimulus_tag_helper/stimulus_controller_builder.rb
@@ -6,9 +6,14 @@ module StimulusTagHelper
       delegate :properties, :aliases, :attribute_method_name_for, :property_method_name_for, to: StimulusTagHelper
     end
 
-    def initialize(identifier:, template:)
+    attr_reader :identifier, :template, :tag, :tag_options
+
+    # all the options are tag options of now
+    def initialize(identifier:, template:, tag: nil, **tag_options)
       @identifier = identifier
       @template = template
+      @tag = tag
+      @tag_options = tag_options
     end
 
     properties.each do |name|
@@ -18,7 +23,7 @@ module StimulusTagHelper
 
       class_eval(<<-RUBY, __FILE__, __LINE__ + 1)
         def #{name}(*args, **kwargs)                                                        # def values(*args, **kwargs)
-          @template.stimulus_#{attribute_method_name}(*args.unshift(@identifier), **kwargs) #   @template.stimulus_values_attributes(*args.unshift(@identifier), **kwargs)
+          template.stimulus_#{attribute_method_name}(*args.unshift(identifier), **kwargs)   #   template.stimulus_values_attributes(*args.unshift(identifier), **kwargs)
         end                                                                                 # end
 
         alias_method :#{alias_name}, :#{name}                                               # alias_method :value, :values
@@ -26,7 +31,7 @@ module StimulusTagHelper
         alias_method :#{alias_name}_attribute, :#{name}                                     # alias_method :value_attribute, :values
 
         def #{property_method_name}(*args, **kwargs)                                        # def values_properties(*args, **kwargs)
-          @template.stimulus_#{property_method_name}(*args.unshift(@identifier), **kwargs)  #   @template.stimulus_values_properties(*args.unshift(@identifier), **kwargs)
+          template.stimulus_#{property_method_name}(*args.unshift(identifier), **kwargs)    #   template.stimulus_values_properties(*args.unshift(identifier), **kwargs)
         end                                                                                 # end
 
         alias_method :#{alias_name}_property, :#{property_method_name}                      # alias_method :value_property, :values_properties
@@ -34,7 +39,19 @@ module StimulusTagHelper
     end
 
     def attributes(**attributes)
-      @template.stimulus_attributes(@identifier, **attributes)
+      template.stimulus_attributes(identifier, **attributes)
+    end
+
+    def properties(**properties)
+      template.stimulus_properties(identifier, **properties)
+    end
+
+    def capture(&block)
+      return template.capture(self, &block) unless tag
+
+      template.stimulus_controller_tag(identifier, tag: tag, **tag_options) do
+        template.capture(self, &block)
+      end
     end
   end
 end